ZiqsyInternal/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.54.0",
    "email-validator>=2.2.0",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "markdown>=3.8.1",
    "openai>=1.88.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.0",
    "pdfkit>=1.0.0",
    "PyMySQL==1.1.0",
    "mysqlclient==2.2.4",
    "sqlalchemy>=2.0.41",
    "werkzeug>=3.1.3",
    "xlrd>=2.0.2",
]
